Move proxy::task_main/handle_client to pglb
